Moving the new block helper & custom renderers
Moving the new block helper (now the block builder) from the Minecraft
class to the Block class so all mods can bask in it's amazingness

Changing the block builder to use an object for it's arguments instead
of actual arguments.

Rewriting most of the minecraft blocks I'd programmed with the old
system, but now with the new, easier, prettier system.

Implemented custom renderers. Kind of.